backend/requirements-local.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.104.1
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
